# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FleetOps SADC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      tags: [Drivers]
      operationId: listDrivers
      summary: List all drivers in the org
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
    post:
      tags: [Drivers]
      operationId: createDriver
      summary: Add a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
  /drivers/{driverId}:
    get:
      tags: [Drivers]
      operationId: getDriver
      summary: Get a single driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
    patch:
      tags: [Drivers]
      operationId: updateDriver
      summary: Update driver details
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
    delete:
      tags: [Drivers]
      operationId: deleteDriver
      summary: Remove a driver
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # ─── Loads ────────────────────────────────────────────────────────────────
...
  /trips/{tripId}/disbursements:
    get:
      tags: [Disbursements]
      operationId: listDisbursements
      summary: List disbursements for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Disbursement'
    post:
      tags: [Disbursements]
      operationId: createDisbursement
      summary: Log a disbursement on a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disbursement'
  /trips/{tripId}/disbursements/{disbursementId}:
    patch:
      tags: [Disbursements]
      operationId: updateDisbursement
      summary: Update a disbursement
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disbursement'
    delete:
      tags: [Disbursements]
      operationId: deleteDisbursement
      summary: Delete a disbursement
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
        - name: disbursementId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # ─── Compliance Documents ─────────────────────────────────────────────────
...
  /fuel-purchases:
    get:
      tags: [Fuel]
      operationId: listFuelPurchases
      summary: List fuel purchases for the org
      parameters:
        - name: truckId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuelPurchase'
    post:
      tags: [Fuel]
      operationId: createFuelPurchase
      summary: Log a fuel purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuelPurchaseInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuelPurchase'
  /fuel-purchases/{purchaseId}:
    patch:
      tags: [Fuel]
      operationId: updateFuelPurchase
      summary: Update a fuel purchase
      parameters:
        - name: purchaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuelPurchaseUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuelPurchase'
    delete:
      tags: [Fuel]
      operationId: deleteFuelPurchase
      summary: Delete a fuel purchase
      parameters:
        - name: purchaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # ─── Border Charge Catalog ────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Driver:
      type: object
      required: [id, orgId, name, status, loadsCompleted, createdAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        phone:
          type: [string, 'null']
        nationality:
          type: [string, 'null']
        passportNo:
          type: [string, 'null']
        status:
          type: string
        loadsCompleted:
          type: integer
        linkedUserId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    DriverInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        nationality:
          type: string
        passportNo:
          type: string
        status:
          type: string
    DriverUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        nationality:
          type: string
        passportNo:
          type: string
        status:
          type: string

    # ─── Load Schemas ─────────────────────────────────────────────────────
...
    Disbursement:
      type: object
      required: [id, tripId, date, country, chargeType, amount, currency, paidBy, createdAt]
      properties:
        id:
          type: string
        tripId:
          type: string
        date:
          type: string
          format: date
        borderPost:
          type: [string, 'null']
        country:
          type: string
        chargeType:
          type: string
        amount:
          type: number
        currency:
          type: string
        paidBy:
          type: string
        receiptUrl:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        amountInBase:
          type: [number, 'null']
        createdAt:
          type: string
          format: date-time
    DisbursementInput:
      type: object
      required: [date, country, chargeType, amount, currency, paidBy]
      properties:
        date:
          type: string
          format: date
        borderPost:
          type: string
        country:
          type: string
        chargeType:
          type: string
        amount:
          type: number
        currency:
          type: string
        paidBy:
          type: string
        receiptUrl:
          type: string
        notes:
          type: string
    DisbursementUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
        borderPost:
          type: string
        country:
          type: string
        chargeType:
          type: string
        amount:
          type: number
        currency:
          type: string
        paidBy:
          type: string
        receiptUrl:
          type: string
        notes:
          type: string

    # ─── Compliance Schemas ────────────────────────────────────────────────
...
    FuelPurchase:
      type: object
      required: [id, orgId, truckId, date, country, litres, pricePerLitre, currency, total, createdAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        truckId:
          type: string
        truckFleetNo:
          type: [string, 'null']
        date:
          type: string
          format: date
        country:
          type: string
        litres:
          type: number
        pricePerLitre:
          type: number
        currency:
          type: string
        total:
          type: number
        createdAt:
          type: string
          format: date-time
    FuelPurchaseInput:
      type: object
      required: [truckId, date, country, litres, pricePerLitre, currency]
      properties:
        truckId:
          type: string
        date:
          type: string
          format: date
        country:
          type: string
        litres:
          type: number
        pricePerLitre:
          type: number
        currency:
          type: string
    FuelPurchaseUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
        country:
          type: string
        litres:
          type: number
        pricePerLitre:
          type: number
        currency:
          type: string

    # ─── Border Catalog Schemas ────────────────────────────────────────────
...