# lib/api-zod/src/generated/types/fuelPurchaseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FleetOps SADC API
 * OpenAPI spec version: 0.1.0
 */

export interface FuelPurchaseUpdate {
  date?: Date;
  country?: string;
  litres?: number;
  pricePerLitre?: number;
  currency?: string;
}